refactor(recon): promote per-host store out of recon_insights (fix C0302)
recon_insights.py had grown to 592 lines (tripping the 500-line C0302
gate) by absorbing all the per-host facet I/O I added. Promote that
layer into tools/recon_host_store.py - the typed reader/writer pairs for
the hosts/<fqdn>/ artefacts (insight / host / notes / tls / ports /
findings). recon_insights keeps validation + the finalise_recon
orchestration that drives those writers, and re-exports the moved names
(via __all__) so existing `from tools.recon_insights import save_insight`
consumers - the Annotate Host tool - keep working unchanged.

Result: recon_insights 419 lines (pylint 10.00, C0302 cleared),
recon_host_store 247. Tests mirror the split: the persistence classes
move to tests/tools/test_recon_host_store.py (importing from the new
module), leaving test_recon_insights.py at 299. Behaviour unchanged;
diff-cover stays at 100%.
